Imports restructurized, renaming: 'GalacticDiskEnum' -> 'GalacticDiskTypes'
