+++ alcor/models/__init__.py
@@ -1,3 +1,4 @@
 from .group import Group
 from .star import (STAR_PARAMETERS_NAMES,
+                   GalacticDiskTypes,
                    Star)
